Fixed some common code warnings/issues

+++ branches/marco/ports/vice/src/c64/tfe.c
@@ -1172,8 +1172,12 @@ void tfe_sideeffects_write_pp(WORD ppaddress, int odd_address)
 
             if(odd_address && (ppaddress == TFE_PP_ADDR_LOG_ADDR_FILTER+6))
                 log_message(tfe_log,"set hash filter: %02x:%02x:%02x:%02x:%02x:%02x",
-                            tfe_hash_mask[0],tfe_hash_mask[1],tfe_hash_mask[2],
-                            tfe_hash_mask[3],tfe_hash_mask[4],tfe_hash_mask[5]);
+                            (unsigned int)tfe_hash_mask[0],
+                            (unsigned int)tfe_hash_mask[1],
+                            (unsigned int)tfe_hash_mask[2],
+                            (unsigned int)tfe_hash_mask[3],
+                            (unsigned int)tfe_hash_mask[4],
+                            (unsigned int)tfe_hash_mask[5]);
         }
         break;
